Wrapping up a first version of the analysis
The major novelty here is the addition of the meta-regression with constituent proportions. The method employs log-ratios as in compositional data analysis.

Descriptive statistics of the composition are added.

Small change in constituent data: all years are kept to have more representative means per city.

Update of the README to present a (very rough) summary of the analysis so far.